Split up packages using lerna.
