+++ packages/ste-core/.npmignore
