+++ packages/ste-core/src/index.ts
@@ -1,0 +1,28 @@
+/*!
+ * Strongly Typed Events for TypeScript - Core
+ * https://github.com/KeesCBakker/StronlyTypedEvents/
+ * http://keestalkstech.com
+ *
+ * Copyright Kees C. Bakker / KeesTalksTech
+ * Released under the MIT license
+ */
+
+import { IEventManagement } from "./management";
+import { ISubscribable } from "./definitions/subscribable";
+import {
+  DispatcherBase,
+  DispatcherWrapper,
+  EventListBase
+} from "./dispatching";
+import { Subscription } from "./subscription";
+import { IBaseEventHandling } from "./definitions/handling";
+
+export {
+  DispatcherBase,
+  DispatcherWrapper,
+  EventListBase,
+  Subscription,
+  IEventManagement,
+  ISubscribable,
+  IBaseEventHandling
+};
