+++ packages/ste-core/src/test/bugs-test.ts
@@ -1,0 +1,9 @@
+import {
+  DispatcherBase,
+  DispatcherWrapper,
+  EventListBase,
+  Subscription
+} from "./../";
+import { expect } from "chai";
+
+describe("Bug fixes", () => {});
