+++ packages/ste-core/tsconfig.json
