+++ packages/ste-events/src/index.ts
@@ -1,0 +1,11 @@
+import { EventDispatcher, EventHandlingBase, EventList } from "./events";
+import { IEventHandler, IEventHandling, IEvent } from "./definitions";
+
+export {
+  EventDispatcher,
+  EventHandlingBase,
+  EventList,
+  IEventHandling,
+  IEvent,
+  IEventHandler
+};
