+++ packages/ste-events/src/test/event-test.ts
@@ -1,8 +1,7 @@
 "use strict";
 
 import { expect } from "chai";
-
-import { EventDispatcher, EventList } from "./../";
+import { EventDispatcher, EventList } from "../";
 
 class Dummy {
   constructor(name: string) {}
