+++ packages/ste-signals/src/index.ts
@@ -1,0 +1,11 @@
+import { SignalDispatcher, SignalHandlingBase, SignalList } from "./signals";
+import { ISignalHandling, ISignal, ISignalHandler } from "./definitions";
+
+export {
+  SignalDispatcher,
+  SignalHandlingBase,
+  SignalList,
+  ISignalHandling,
+  ISignal,
+  ISignalHandler
+};
