+++ packages/ste-simple-events/src/index.ts
@@ -1,0 +1,19 @@
+import {
+  SimpleEventDispatcher,
+  SimpleEventHandlingBase,
+  SimpleEventList
+} from "./simple-events";
+import {
+  ISimpleEventHandling,
+  ISimpleEvent,
+  ISimpleEventHandler
+} from "./definitions";
+
+export {
+  SimpleEventDispatcher,
+  SimpleEventHandlingBase,
+  SimpleEventList,
+  ISimpleEventHandling,
+  ISimpleEvent,
+  ISimpleEventHandler
+};
