+++ packages/ste-simple-events/src/test/simple-event-test.ts
