+++ packages/strongly-typed-events/.npmignore
@@ -50,4 +50,10 @@ jspm_packages
 
 # Optional REPL history
 .node_repl_history
-dist
+
+documentation
+examples
+src
+.travis.yml
+tsconfig.json
+dist/test
