+++ packages/strongly-typed-events/src/index.ts
@@ -1,0 +1,73 @@
+/*!
+ * Strongly Typed Events for TypeScript
+ * https://github.com/KeesCBakker/StronlyTypedEvents/
+ * http://keestalkstech.com
+ *
+ * Copyright Kees C. Bakker / KeesTalksTech
+ * Released under the MIT license
+ */
+
+import {
+  DispatcherBase,
+  DispatcherWrapper,
+  EventListBase,
+  Subscription,
+  IEventManagement,
+  ISubscribable,
+  IBaseEventHandling
+} from "ste-core";
+
+import {
+  EventDispatcher,
+  EventHandlingBase,
+  EventList,
+  IEventHandling,
+  IEvent,
+  IEventHandler
+} from "ste-events";
+
+import {
+  SimpleEventDispatcher,
+  SimpleEventHandlingBase,
+  SimpleEventList,
+  ISimpleEventHandling,
+  ISimpleEvent,
+  ISimpleEventHandler
+} from "ste-simple-events";
+
+import {
+  SignalDispatcher,
+  SignalHandlingBase,
+  SignalList,
+  ISignalHandling,
+  ISignal,
+  ISignalHandler
+} from "ste-signals";
+
+export {
+  DispatcherBase,
+  DispatcherWrapper,
+  EventListBase,
+  Subscription,
+  IEventManagement,
+  ISubscribable,
+  IBaseEventHandling,
+  EventDispatcher,
+  EventHandlingBase,
+  EventList,
+  IEventHandling,
+  IEvent,
+  IEventHandler,
+  SimpleEventDispatcher,
+  SimpleEventHandlingBase,
+  SimpleEventList,
+  ISimpleEventHandling,
+  ISimpleEvent,
+  ISimpleEventHandler,
+  SignalDispatcher,
+  SignalHandlingBase,
+  SignalList,
+  ISignalHandling,
+  ISignal,
+  ISignalHandler
+};
