+++ packages/strongly-typed-events/src/test/bugs-test.ts
