+++ packages/strongly-typed-events/src/test/classic-test.ts
