+++ packages/strongly-typed-events/src/test/dispatcher-test.ts
