--- src/index.ts
@@ -1,59 +1,0 @@
-/*!
- * Strongly Typed Events for TypeScript
- * https://github.com/KeesCBakker/StronlyTypedEvents/
- * http://keestalkstech.com
- *
- * Copyright Kees C. Bakker / KeesTalksTech
- * Released under the MIT license
- */
-
-import {
-  IEventHandler,
-  ISignalHandler,
-  ISimpleEventHandler
-} from "./definitions/handlers";
-export { IEventHandler, ISignalHandler, ISimpleEventHandler };
-
-import {
-  IBaseEventHandling,
-  IEventHandling,
-  ISignalHandling,
-  ISimpleEventHandling
-} from "./definitions/handling";
-export {
-  IBaseEventHandling,
-  IEventHandling,
-  ISignalHandling,
-  ISimpleEventHandling
-};
-
-import {
-  IEvent,
-  ISignal,
-  ISimpleEvent,
-  ISubscribable
-} from "./definitions/subscribables";
-export { IEvent, ISignal, ISimpleEvent, ISubscribable };
-
-import {
-  DispatcherBase,
-  DispatcherWrapper,
-  EventListBase
-} from "./dispatching";
-export { DispatcherBase, DispatcherWrapper, EventListBase };
-
-import { Subscription } from "./subscription";
-export { Subscription };
-
-import { EventDispatcher, EventHandlingBase, EventList } from "./events";
-export { EventDispatcher, EventHandlingBase, EventList };
-
-import {
-  SimpleEventDispatcher,
-  SimpleEventHandlingBase,
-  SimpleEventList
-} from "./simple-events";
-export { SimpleEventDispatcher, SimpleEventHandlingBase, SimpleEventList };
-
-import { SignalDispatcher, SignalHandlingBase, SignalList } from "./signals";
-export { SignalDispatcher, SignalHandlingBase, SignalList };
